Wrap some computations into a useMemo
